Significantly rewrite TGBOX-CLI source
The strange CLI program flow was settled when i firstly
encountered the 'click' library. Now, when i learned
more of this beautiful Lib i decided to completely
rewrite the way the commands interact.

Now, main context is defined at the top-level
'cli()' function. Here we process the Session,
Decrypted(Remote/Local)Box, Account, etc and
save it in the context (the click's CTX). This
context is passed to the all child commands,
so we can easily interact with it.

Every command now can check the CTX for the
available fields, and if it's not have some
requirements, -- echo error message and
raise the CheckCTXFailed error.

Also, now DecryptedRemoteBox (drb) and the
TelegramClient (account) fields of the context
Objects (ctx.obj.*) is lazy by default. That
means that there will be no any internet
interaction until drb/account will be
requested by some of the commands.

Also also, i did some updates to the Session
files. There is new Session class on the
new 'session' module, which makes work with
the session files a lot cleaner. I changed
names of some Session fields, so your old
session files will stop working. Please,
make a new State with just cli-init and
the box-open / account-connect

& ETC!

The users-only will not even see any difference,
however, now code is cleaner and more readable (
at least i hope so), this should help in the
future development of the TGBOX-CLI.

This significant rewrite may break some commands,
so now we will need to test it, then it can be
finally released as v1.0. Long way from home!