Disable broken colored output in Pager on Windows
I can not find a way to show colors in the default
Windows CMD Pager (like in 'file-search' or 'help'),
so now it will be not forced (up to Click library
to decide, color or not). You can force by adding
the "TGBOX_CLI_FORCE_PAGER_COLOR" to Env Vars,
although output will be probably broken.

Windows is some piece of fucking shit. I fucking
hate it. Who ever will want to write code on this
bullshit OS? That's absolutely fucking annoying.

All i can recommend is just to use a Cygwin. Don't
want to spend more time on this awful system.