NonProjects/tgbox@d264297 and lower --max-bytes
This commit adds new --use-slow-upload and
--use-slow-download flags to commands.

Also, amount of bytes downloaded/uploaded
at the same time was lowered from 500MB
to 200MB because Telegram is now more
aggressive to parallel interactions
with server. It means that if you want
to upload to your Box 10 files with 100MB
in size then CLI will upload 2 files at
the same time. Before it was 5. You are
free to override this with --max-bytes
and --max-workers options on commands.