Heavily impove text color system in TGBOX-CLI
This commit completely changes how we color text.

The old color codes was full color names in
uppercase, e.g [BRIGHT_BLACK]Text[BRIGHT_BLACK],
by default was *always* a bold text and...
that's it.

Started from now, color codes are defined as
one letter and one digit, followed by optional
formatting options. See this:

B0 is a regular Blue color. Here, "B" means
"Blue" and "0" means that it's regular. The
"B1" is *bright* blue. Currently, only "0"
and "1" supported as second symbol in code.

All codes are:

AVAILABLE_COLORS = {
 'B0': 'blue',
 'B1': 'bright_blue',
 'C0': 'cyan',
 'C1': 'bright_cyan',
 'G0': 'green',
 'G1': 'bright_green',
 'M0': 'magenta',
 'M1': 'bright_magenta',
 'R0': 'red',
 'R1': 'bright_red',
 'W0': 'white',
 'W1': 'bright_white',
 'X0': 'black',
 'X1': 'bright_black',
 'Y0': 'yellow',
 'Y1': 'bright_yellow'
}

Optional formatting is appended to the end
of the Color Code. Available formatting is:

_color_formattings = {
 'i': 'italic',
 'b': 'bold',
 'd': 'dim',
 'u': 'underline',
 'o': 'overline',
 'r': 'reverse',
 'l': 'blink',
 'g': 'bg',
 's': 'strikethrough'
}

Formattings are *always* lowercase, otherwise
ignored. For example, "W1bi" is the bold and
italic, bright white text. "W1bil" formatted
text will blink in Terminal. You can apply
to Color Code any amount of formattings.

Previously, colored string was required to
start *and* end with the code, as you may
see in the start of this text with the
"[BRIGHT_BLACK]" example. Now, to end
color you would need to just use "[X]"

For example,

(
 '[G1bl]This part is bold Green and will blink[X] '
 '[R0o]And this is regular Red with overline[X]'
)

Old `tgbox.tools.color` function, which convert
Color Codes to actual ANSI was renamed, and now
it is `tgbox.tools.colorized`.

That's it with new little colorize() engine! :)

------------------------------------------------

Most changes in this commit is migration to the
new Color Codes. Let's hope that it will work
nicely, otherwise more commits to come.

------------------------------------------------

Also, try Rust in Peace by Megadeth!