Completely revamped TGBOX-CLI tree structure
This commit revamps CLI tree structure, or, to be honest,
adds it. Prior to this commit everything was in one file,
and it was bullshit. This file grew tremendously over
the last years, and I just didn't had enough time to
clean up all this mess. I hope this structure is
better and more clean now.

I'm not currently sure about stability, as I moved
around literally everything. Will need some time
to test it out.